[feat] add utils and improve some files
