+++ src/interfaces/team.ts
@@ -1,6 +1,8 @@
 import { Time } from '@/types/time';
 import { Locale } from '@/types/locale';
 
+export type Gender = 'Masculino' | 'Feminino';
+
 export enum Weekday {
   sunday = 0,
   monday = 1,
@@ -14,7 +16,7 @@ export enum Weekday {
 export default interface ITeam {
   id: string;
   name: string;
-  gender: 'Masculino' | 'Feminino';
+  gender: Gender;
   day: Array<{
     weekday: Weekday;
     startTime: Time;
