+++ src/utils/format-day.ts
@@ -1,0 +1,10 @@
+const formatDay = (date: Date) => {
+  const newDate = new Date(date);
+
+  const day = newDate.getDate();
+  const month = newDate.getMonth() + 1;
+  const year = newDate.getFullYear();
+  return `${day < 10 ? `0${day}` : day}/${month < 10 ? `0${month}` : month}/${year}`;
+};
+
+export default formatDay;
