+++ src/utils/format-time.ts
@@ -1,0 +1,7 @@
+const formatTime = (hours: number, minutes: number) => {
+  const formattedHours = hours < 10 ? `0${hours}` : hours;
+  const formattedMinutes = minutes < 10 ? `0${minutes}` : minutes;
+  return `${formattedHours}:${formattedMinutes}`;
+};
+
+export default formatTime;
